Began Restructuring for Other ROS Versions
Refactored the websites back end to put all things related to ROS 2 in its own file.

TODO: figure out how to add another layer of indentation to our primary sidebar. This will clean it up, so the user can expand specifically the section they are using. Also, refactor media files to be stored in the same directory as the distribution they represent.
Co-Authored-By: Luca Vitti <56119921+LucaVits@users.noreply.github.com>